requirements.txt
discord.py==1.7.3
sentry-sdk==1.3.1
tortoise-orm==0.17.7
asyncpg==0.24.0
python-dotenv==0.19.0
aerich==0.5.8
